dbGet AMICommandFailure when Database entry not found
Change manager.py to handle a deferred function for the dbGet
method when a "Resopnse: Error" is returned by AMI.  This
returns a None value on the dbGet callback.

This fixes an issue I had where an "Response: Error" to getDb
would raise an error.AMICommandFailure(message) exception when
the errorUnlessResponse method was called.

Also add a call deregisterEvent("DBGetResponse", extractValue) prior
to the dbGet callbacks.