Correct fastagi.controlStreamFile() for Asterisk 12.
Asterisk 12 changes the response of AGI 'control stream file' to include the
endpos of where the playback stopped.

This patch changes the response of fastagi.controlStreamFile() from being a
single string to being a tuple of (result, endpos), which makes it consistent
with getOption(), streamFile(), and recordFile().

For Asterisk 11 and earlier, the response is (result, None).