starpy/manager: Allow AMIFactories to reconnect
This patch makes the AMIFactory object inherit from the twisted
ReconnectingClientFactory. This will automatically attempt to reconnect to
Asterisk if the connection is broken. Additionally, the constructor for
the AMIFactory now takes in an optional parameter, on_reconnect. If
the instantiator passes in a function to this parameter, it will be called
with the new login deferred when a re-connect attempt is made.