AMI autoconnect after failed connection
- In manager.AMIFactory added ability to reconnect after connection
loss (as in parent class protocol.ReconnectingClientFactory)
- To test such activity a class ChannelTracker from
examples/priexhaustionbare.py was used. As a result were fixed some bugs
and code was cleaned up a bit in aforementioned file and in
utilapplication.py, which contains class AMISpecifier (which was fixed
too).

So, after applying changes in this commit, it will be possible to
auto-connect to Asterisk server even after connection was lost for
relatively long time (from several seconds and up).