Add support for Python 3
manager: Replace 'async' reference with 'nowait'.  This change breaks
backwards compatibility with previous versions of the testsuite,
requiring a matching change there. 'async' is now a reserved word
in python and can not be used.

manager and fastagi: twisted in general uses bytes for passing data back
and forth while internally we generally want to deal with strings.  This
requires converting between the two as we interact with the twisted
library.

ASTERISK-26826

Change-Id: I2fab42abb5318a5db2af3bcd291fc2db95b87e15